group Management latest version

+++ src/views/manager/group/GroupIndex.vue
@@ -165,7 +165,7 @@
                         </b-button>
                         <b-modal size="xl" centered  hide-footer id="addManager" ref="addManager" title="配置管理员">
                             <div class="pb-5">
-                                <b-button :size="template_size" variant="outline-primary" class="mb-3 offset-10"
+                                <b-button :size="template_size" variant="outline-primary" class="mb-3 offset-10 styledBtn fontedBtn" style="margin-left: 82% !important;"
                                           @click="()=>{newManager = true; editManager = false; resetManager = false; new_Manager = {name:'',description:'',password:null}}">
                                     新增管理员
                                 </b-button>
@@ -180,11 +180,11 @@
                                     </template>
                                     <template slot="action" slot-scope="row">
                                         <b-button-group>
-                                            <b-button :key="'edit' + row.id" :size="template_size" variant="outline-primary"
+                                            <b-button class="styledBtn" :key="'edit' + row.id" :size="template_size" variant="outline-primary"
                                                       @click="()=>{editManager = true; newManager = false; resetManager = false; edit_Manager = {id:row.item.id,description:row.item.description}}">
                                                 修改信息
                                             </b-button>
-                                            <b-button :key="'reset' + row.id" :size="template_size"
+                                            <b-button class="styledBtn" :key="'reset' + row.id" :size="template_size"
                                                       variant="outline-primary"
                                                       @click="()=>{resetManager = true; newManager = false; editManager = false; reset_Manager = {id:row.item.id,password:''}}">
                                                 重置密码
@@ -192,74 +192,86 @@
                                         </b-button-group>
                                     </template>
                                 </b-table>
-                                <b-form @submit="newManagerSave" class="container w-25 pt-3" v-if="newManager">
-                                    <b-form-group id="input-group-7" label-for="name">
-                                        <b-form-input
-                                                v-model="new_Manager.name"
-                                                required
-                                                autocomplete="username"
-                                                placeholder="管理员名称"
-                                        ></b-form-input>
-                                    </b-form-group>
-                                    <b-form-group id="input-group-8" label-for="input-2">
-                                        <b-form-textarea
-                                                rows="3"
-                                                no-resize
-                                                v-model="new_Manager.description"
-                                                required
-                                                placeholder="备注"
-                                        ></b-form-textarea>
-                                    </b-form-group>
-                                    <b-form-group id="input-group-9" label-for="input-2">
-                                        <b-form-input
-                                                v-model="new_Manager.password"
-                                                required
-                                                autocomplete="new-password"
-                                                type="password"
-                                                placeholder="密码"
-                                        ></b-form-input>
-                                    </b-form-group>
-                                    <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
-                                    </b-button>
-                                    <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
-                                              @click="()=>{newManager = false; new_Manager = {name:'',description:'',password:null}}">取 消
-                                    </b-button>
-                                </b-form>
-                                <b-form @submit="editManagerSave" class="container w-25 pt-3" v-if="editManager">
-                                    <b-form-group id="input-group-10" label-for="input-2">
-                                        <b-form-textarea
-                                                id="edit_Manager_description"
-                                                rows="3"
-                                                no-resize
-                                                v-model="edit_Manager.description"
-                                                required
-                                                placeholder="备注"
-                                        ></b-form-textarea>
-                                    </b-form-group>
-                                    <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
-                                    </b-button>
-                                    <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
-                                              @click="()=>{editManager = false; edit_Manager = {id: null,description:''}}">取
-                                        消
-                                    </b-button>
-                                </b-form>
-                                <b-form @submit="resetManagerSave" class="container w-25 pt-3" v-if="resetManager">
-                                    <b-form-group id="input-group-11" label-for="input-2">
-                                        <b-form-input
-                                                v-model="reset_Manager.password"
-                                                required
-                                                autocomplete="new-password"
-                                                type="password"
-                                                placeholder="密码"
-                                        ></b-form-input>
-                                    </b-form-group>
-                                    <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
-                                    </b-button>
-                                    <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
-                                              @click="()=>{resetManager = false; reset_Manager = {id: null,password:''}}">取
-                                        消
-                                    </b-button>
-                                </b-form>
+                                <b-modal hide-footer centered  v-model="newManager" title="修改集群">
+                                    <div>
+                                        <b-form @submit="newManagerSave" class="container pt-3">
+                                            <b-form-group id="input-group-7" label-for="name">
+                                                <b-form-input
+                                                        v-model="new_Manager.name"
+                                                        required
+                                                        autocomplete="username"
+                                                        placeholder="管理员名称"
+                                                ></b-form-input>
+                                            </b-form-group>
+                                            <b-form-group id="input-group-8" label-for="input-2">
+                                                <b-form-textarea
+                                                        rows="3"
+                                                        no-resize
+                                                        v-model="new_Manager.description"
+                                                        required
+                                                        placeholder="备注"
+                                                ></b-form-textarea>
+                                            </b-form-group>
+                                            <b-form-group id="input-group-9" label-for="input-2">
+                                                <b-form-input
+                                                        v-model="new_Manager.password"
+                                                        required
+                                                        autocomplete="new-password"
+                                                        type="password"
+                                                        placeholder="密码"
+                                                ></b-form-input>
+                                            </b-form-group>
+                                            <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
+                                            </b-button>
+                                            <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
+                                                      @click="()=>{newManager = false; new_Manager = {name:'',description:'',password:null}}">取 消
+                                            </b-button>
+                                        </b-form>
+                                    </div>
+                                </b-modal>
+                                <b-modal hide-footer centered  v-model="editManager" title="修改集群">
+                                    <div>
+                                        <b-form @submit="editManagerSave" class="container pt-3" >
+                                            <b-form-group id="input-group-10" label-for="input-2">
+                                                <b-form-textarea
+                                                        id="edit_Manager_description"
+                                                        rows="3"
+                                                        no-resize
+                                                        v-model="edit_Manager.description"
+                                                        required
+                                                        placeholder="备注"
+                                                ></b-form-textarea>
+                                            </b-form-group>
+                                            <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
+                                            </b-button>
+                                            <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
+                                                      @click="()=>{editManager = false; edit_Manager = {id: null,description:''}}">取
+                                                消
+                                            </b-button>
+                                        </b-form>
+                                    </div>
+                                </b-modal>
+                                <b-modal hide-footer centered  v-model="resetManager" title="修改集群">
+                                    <div>
+                                        <b-form @submit="resetManagerSave" class="container pt-3">
+                                            <b-form-group id="input-group-11" label-for="input-2">
+                                                <b-form-input
+                                                        v-model="reset_Manager.password"
+                                                        required
+                                                        autocomplete="new-password"
+                                                        type="password"
+                                                        placeholder="密码"
+                                                ></b-form-input>
+                                            </b-form-group>
+                                            <b-button class="mt-3 my-4 col-5 float-left" block type="submit" variant="primary">保 存
+                                            </b-button>
+                                            <b-button class="mt-3 my-4 col-5 float-right" block variant="primary"
+                                                      @click="()=>{resetManager = false; reset_Manager = {id: null,password:''}}">取
+                                                消
+                                            </b-button>
+                                        </b-form>
+                                    </div>
+                                </b-modal>
                             </div>
                         </b-modal>
                     </b-button-group>
@@ -438,7 +450,6 @@
                     return
                 }
                 else {
-                    this.$refs['newGroup'].hide();
                     this.run();
                     GroupService
                         .create(this.newGroup)
@@ -458,6 +469,7 @@
                                         this.allgroup.list = data.results;
                                         this.allgroup.total = data.paging.count;
                                         this.$emit("data-ready");
+                                        this.$refs['newGroup'].hide();
                                     })
                                     .catch(() => {
                                         this.$emit("data-failed");
@@ -500,35 +512,37 @@
                     });
             },
             deleteGroup() {
-                this.run();
-                GroupService
-                    .deleteGroups({ids: JSON.stringify(this.selected)})
-                    .then(res => {
-                        if (res.results === 'success')
-                            GroupService
-                                .fetchList({...this.queryParam, ...this.queryDebounceParam})
-                                .then(data => {
-                                    data.results.forEach(item => {
-                                        if (item.checked === undefined) {
-                                            item.checked = false;
-                                        }
-                                        if (item.locked === undefined) {
-                                            item.locked = false;
-                                        }
+                if (confirm("您确定要删除该集群吗？")){
+                    this.run();
+                    GroupService
+                        .deleteGroups({ids: JSON.stringify(this.selected)})
+                        .then(res => {
+                            if (res.results === 'success')
+                                GroupService
+                                    .fetchList({...this.queryParam, ...this.queryDebounceParam})
+                                    .then(data => {
+                                        data.results.forEach(item => {
+                                            if (item.checked === undefined) {
+                                                item.checked = false;
+                                            }
+                                            if (item.locked === undefined) {
+                                                item.locked = false;
+                                            }
+                                        });
+                                        this.allgroup.list = data.results;
+                                        this.allgroup.total = data.paging.count;
+                                        this.$emit("data-ready");
+                                    })
+                                    .catch(() => {
+                                        this.$emit("data-failed");
                                     });
-                                    this.allgroup.list = data.results;
-                                    this.allgroup.total = data.paging.count;
-                                    this.$emit("data-ready");
-                                })
-                                .catch(() => {
-                                    this.$emit("data-failed");
-                                });
-                        else
+                            else
+                                this.$emit("data-failed");
+                        })
+                        .catch(() => {
                             this.$emit("data-failed");
-                    })
-                    .catch(() => {
-                        this.$emit("data-failed");
-                    });
+                        });
+                }
             },
             editOpen(row) {
                 this.editItem.id = row.item.id;
@@ -545,7 +559,6 @@
             },
             updateGroup(evt) {
                 evt.preventDefault();
-                this.$refs['editGroup'].hide();
                 this.run();
                 GroupService
                     .update(this.editItem)
@@ -565,10 +578,15 @@
                                     this.allgroup.list = data.results;
                                     this.allgroup.total = data.paging.count;
                                     this.$emit("data-ready");
+                                    this.$refs['editGroup'].hide();
                                 })
                                 .catch(() => {
                                     this.$emit("data-failed");
                                 });
+                        else if (res.results === 'nameError') {
+                            alert("该集群名已存在。");
+                            this.$emit("data-failed");
+                        }
                         else
                             this.$emit("data-failed");
                     })
@@ -578,7 +596,6 @@
             },
             newManagerSave(evt) {
                 evt.preventDefault();
-                this.newManager = false;
                 this.run();
                 GroupService
                     .addManager({groupID: this.Managers.groupID, data: this.new_Manager})
@@ -600,10 +617,15 @@
                                     let selectedData = data.results.filter(obj => { return obj.id === this.Managers.groupID});
                                     this.Managers.list = selectedData[0].groupManagers;
                                     this.$emit("data-ready");
+                                    this.newManager = false;
                                 })
                                 .catch(() => {
                                     this.$emit("data-failed");
                                 });
+                        else if (res.results === 'managerNameError') {
+                            alert("该账号已存在。");
+                            this.$emit("data-failed");
+                        }
                         else
                             this.$emit("data-failed");
                     })
